requirements.txt
asn1crypto==1.3.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
Flask==1.1.1
Flask-Cors==3.0.8
google-api-core==1.16.0
google-auth==1.12.0
google-cloud-speech==1.3.2
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
nltk==3.4.5
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.23.1
pyOpenSSL==19.1.0
PySocks==1.7.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
